refactor: info in policyreport package (#4598)
